Fixed "[muted]" when changing unmuted volume.
This bug occured with me, when I was just changing the volume, pasystray
reported the volume as muted, despite that not being the case. This is
fixed by XOR-ing a new variable called "is_muting" with the mute status.
Also used the mute callback (they were previously unused, as far as I
could tell, but were added in).