Disable status icon tooltip for connected state
* Provide `systray_impl_set_has_tooltip` to allow disabling the tooltip.
* Provide a new option (`no-icon-tooltip`) and setting (`icon_tooltip`,
  the inverse of the option), to disable the status icon tooltip for the
  connected state. Some systray implementations display a temporary
  tooltip on startup; that still happens. Perhaps in the future the
  status icon tooltip can be updated for other (failure) states.
* Initialize `menu_infos_t->settings` before calling `systray_create`.
  Not necessary for this patch but, I think, still a good idea.

This closes #99. The status icon tooltip for the connected state
interferes with the notifications generated by the scroll callback. The
menu tooltips don't interfere, though even if they did the menu already
clearly displays the current volume. The status icon tooltips for other
states do not generate notifications, so should still be displayed.