Use volume max for icon instead of avg
Once again, given that pasystray will stop increasing the volume once
one of the channels arrived to max, it seems logical to index the volume
icon on the maximum value of all channels.

With avg, if one of the channels is set a lot higher than the other
ones, the icon could show a low volume while the maximum is reached.
When using channels volume as a balance setting for instance this is
misguiding as one could think there is some volume left to increase when
in fact it is set to the maximum.