Force X11 if built without AppIndicator
When built without AppIndicator, this is by definition an X11 tray icon.
When built with GTK3 (which is usually built to support both Wayland and
X11), and run on a Wayland desktop, X11/XWayland is the only viable
option for this program, and needs to be forced to start properly.