Fix build with GCC 15
C23 changes the implication of a function declaration without arguments
from an undetermined number and type to `void` (no arguments).  While that
is true in most of the cases here, init() and destroy() actually take a
settings_t*.  This defines the arguments (or lack thereof) of all
functions, avoiding any ambiguity.

https://gcc.gnu.org/gcc-15/porting_to.html